Add transfer toggle function and comment it temporarily

+++ constants/root/contract.constant.ts
@@ -1,9 +1,10 @@
-import { ApiPromise, WsProvider } from '@polkadot/api';
+import { ApiPromise, Keyring, WsProvider } from '@polkadot/api';
 import { getApiOptions } from '@therootnetwork/api';
 import { LogError } from '../../utils/common.util';
 import { Module } from '../../enums/common.enum';
 import { Signer } from '@futureverse/signer';
 import { AssetRegister } from '@futureverse/asset-register/v2'
+import { KeyringPair } from '@polkadot/keyring/types';
 
 export const ROOT_NETWORK_WS_URL = process.env.NEXT_PUBLIC_ROOT_NETWORK_WS_URL;
 
@@ -25,31 +26,36 @@ export let ASSET_REGISTER_SDK: AssetRegister;
 
 export let API: ApiPromise;
 export let SIGNER: Signer;
+export let KEYRING_SIGNER: KeyringPair;
 
 
-export async function InitializeContractEssentialData(_signer: Signer) {
+export async function InitializeContractEssentialData(_signer?: Signer, _keyringSigner?: KeyringPair) {
 
-    if (API) return LogError(Module.RootContractConstant, 'Attempted to initialize in Singleton Pattern. Api already initialized');
-    API = await ApiPromise.create({ ...getApiOptions(), provider: PROVIDER });
+  if (API) return LogError(Module.RootContractConstant, 'Attempted to initialize in Singleton Pattern. Api already initialized');
+  API = await ApiPromise.create({ ...getApiOptions(), provider: PROVIDER });
 
-    if (SIGNER) return LogError(Module.RootContractConstant, 'Attempted to initialize in Singleton Pattern. Signer already initialized');
-    SIGNER = _signer;
+  if(KEYRING_SIGNER) return LogError(Module.RootContractConstant, 'Attempted to initialize in Singleton Pattern. Keyring Signer already initialized');
+  if(_keyringSigner) KEYRING_SIGNER = _keyringSigner;
 
-    if (ASSET_REGISTER_SDK) return LogError(Module.RootContractConstant, 'Attempted to initialize in Singleton Pattern. Asset Register SDK already initialized');
-    ASSET_REGISTER_SDK = new AssetRegister({
-        url: ROOT_GQL_API_URL,
-        auth: {
-          sign: async (message) => {
-            return await SIGNER.signMessage({ raw: message as `0x${string}` });
-          },
-          storage: {
-            set: (key: string, value: string) => localStorage.setItem(key, value),
-            get: (key: string) => localStorage.getItem(key),
-          },
-          domain: DOMAIN,
-          origin: ORIGIN,
-          chainId: Number(CHAIN_ID),
-          walletAddress: await SIGNER.getAddress() as `0x${string}`,
-        },
-      });
+  if (SIGNER) return LogError(Module.RootContractConstant, 'Attempted to initialize in Singleton Pattern. Signer already initialized');
+  if (_signer) SIGNER = _signer;
+  
+
+  if (ASSET_REGISTER_SDK) return LogError(Module.RootContractConstant, 'Attempted to initialize in Singleton Pattern. Asset Register SDK already initialized');
+  ASSET_REGISTER_SDK = new AssetRegister({
+    url: ROOT_GQL_API_URL,
+    auth: {
+      sign: async (message) => {
+        return await SIGNER.signMessage({ raw: message as `0x${string}` });
+      },
+      storage: {
+        set: (key: string, value: string) => localStorage.setItem(key, value),
+        get: (key: string) => localStorage.getItem(key),
+      },
+      domain: DOMAIN,
+      origin: ORIGIN,
+      chainId: Number(CHAIN_ID),
+      walletAddress: await SIGNER.getAddress() as `0x${string}`,
+    },
+  });
 }
